Refactor CollectibleAsset component structure by separating skeleton and utility functions
- Introduced CollectibleAssetSkeleton as a standalone component for better modularity.
- Added utility functions for file type detection (image, HTML, video, 3D model) in a new utils.ts file to enhance code organization and reusability.
