+++ sdk/src/react/ui/components/collectible-card/CollectibleCard.tsx
@@ -11,8 +11,8 @@ import type {
 import { useCurrency } from '../../../hooks';
 import { ActionButton } from '../_internals/action-button/ActionButton';
 import { CollectibleCardAction } from '../_internals/action-button/types';
-import { CollectibleAsset } from './CollectibleAsset';
 import { Footer } from './Footer';
+import { CollectibleAsset } from './collectible-asset/CollectibleAsset';
 
 function CollectibleSkeleton() {
 	return (
@@ -145,7 +145,7 @@ export function CollectibleCard({
 					/>
 
 					{(highestOffer || collectible) && !balanceIsLoading && (
-						<div className="-bottom-action-offset absolute flex w-full items-center justify-center bg-overlay-light p-2 backdrop-blur transition-transform duration-200 ease-in-out group-hover:translate-y-[-44px]">
+						<div className="-bottom-16 absolute flex w-full origin-bottom items-center justify-center bg-overlay-light p-2 backdrop-blur transition-transform duration-200 ease-in-out group-hover:translate-y-[-64px]">
 							<ActionButton
 								chainId={chainId}
 								collectionAddress={collectionAddress}
